Migrate nom to 5
Warning: This migration also changes the parsing logic. The main
difference to the previous implementation is how strings with escaped
characters are passed.

Previously, we eat the input byte by byte, accumulating into a Vec<u8>
in the background.

Now, we _first_ try to eat an escaped character. If successful, it is
added to the accumulator and we eat further. If not successful, we try
the _second_ alternative: The escape character itself (backslash). If
this fails as well, we resort to our final alternative: Eating a _slice_
until we encounter an escape character or a terminator.

The parsing result is the same on my corpus of Indentex documents, but
still there might be a corner case somewhere.